pyproject.toml
[tool.poetry]
name = "tq"
version = "0.1.0"
description = ""
authors = ["sliverKi <lovee2756@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
jwt = "^1.3.1"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
dotenv = "^0.0.5"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
